requirements.txt
flet>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
geopy>=2.3.0
plyer>=2.1.0
